separated API trotterisation from operations
which has the below benefits:
- the remaining functions in operations.cpp are precise and do not need to be user-configured for accuracy (i.e. no Trotter hyperparameters)
- the remaining functions in operations.cpp merely call the backend and do not include any bespoke logic (i.e. Trotter circuit scheduling)
- incoming new Trotter functions for dynamical simulation will be more clearly delineated from the "standard" (and relatively boring) operations
- the Trotter logic is isolated in preparation for it becoming more substantial with the introduction of randomisation, commuting groups, and that necessary for Linblad master equation solving
