+++ quest/src/api/CMakeLists.txt
@@ -12,5 +12,6 @@ target_sources(QuEST
   operations.cpp
   paulis.cpp
   qureg.cpp
+  trotterisation.cpp
   types.cpp
 )
